move generator to its own file
