+++ .cargo/config.toml
@@ -1,2 +1,3 @@
 [alias]
 db = ["run", "--bin", "db"]
+generate = ["run", "--bin", "generate"]
